[Refactor] quest.cpp/h からquest-definition.cpp/h を分離した
QuestType には定義データと動的記録の両方が含まれているが、中長期的にQuestDefinition とQuestRecord に分離する予定
