+++ VisualStudio/Hengband/Hengband.vcxproj.filters
@@ -2595,6 +2595,9 @@
     <ClCompile Include="..\..\src\io\temp-file.cpp">
       <Filter>io</Filter>
     </ClCompile>
+    <ClCompile Include="..\..\src\system\dungeon\quest-definition.cpp">
+      <Filter>system\dungeon</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\..\src\combat\shoot.h">
@@ -5627,6 +5630,9 @@
     </ClInclude>
     <ClInclude Include="..\..\src\io\temp-file.h">
       <Filter>io</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\src\system\dungeon\quest-definition.h">
+      <Filter>system\dungeon</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
