+++ src/core/game-play.cpp
@@ -25,7 +25,6 @@
 #include "core/visuals-reseter.h"
 #include "core/window-redrawer.h"
 #include "dungeon/dungeon-processor.h"
-#include "dungeon/quest.h"
 #include "floor/floor-changer.h"
 #include "floor/floor-leaver.h"
 #include "floor/floor-mode-changer.h"
@@ -73,6 +72,7 @@
 #include "sv-definition/sv-weapon-types.h"
 #include "system/angband-system.h"
 #include "system/angband-version.h"
+#include "system/dungeon/quest-definition.h"
 #include "system/enums/monrace/monrace-id.h"
 #include "system/floor/floor-info.h"
 #include "system/floor/wilderness-grid.h"
