+++ src/dungeon/quest-completion-checker.cpp
@@ -4,6 +4,7 @@
 #include "floor/floor-util.h"
 #include "grid/grid.h"
 #include "object-enchant/item-apply-magic.h"
+#include "system/dungeon/quest-definition.h"
 #include "system/enums/terrain/terrain-tag.h"
 #include "system/floor/floor-info.h"
 #include "system/grid-type-definition.h"
