+++ src/dungeon/quest-monster-placer.cpp
@@ -1,9 +1,9 @@
 #include "dungeon/quest-monster-placer.h"
-#include "dungeon/quest.h"
 #include "floor/floor-generator-util.h"
 #include "monster-floor/monster-generator.h"
 #include "monster-floor/place-monster-types.h"
 #include "monster/monster-info.h"
+#include "system/dungeon/quest-definition.h"
 #include "system/floor/floor-info.h"
 #include "system/grid-type-definition.h"
 #include "system/monrace/monrace-definition.h"
