+++ src/floor/floor-events.cpp
@@ -25,6 +25,7 @@
 #include "sv-definition/sv-ring-types.h"
 #include "system/angband-system.h"
 #include "system/dungeon/dungeon-definition.h"
+#include "system/dungeon/quest-definition.h"
 #include "system/floor/floor-info.h"
 #include "system/grid-type-definition.h"
 #include "system/item/item-entity.h"
