+++ src/floor/floor-leaver.cpp
@@ -1,5 +1,4 @@
 #include "floor/floor-leaver.h"
-#include "dungeon/quest.h"
 #include "floor/floor-mode-changer.h"
 #include "floor/floor-save-util.h"
 #include "floor/floor-save.h"
@@ -16,6 +15,7 @@
 #include "save/floor-writer.h"
 #include "spell-class/spells-mirror-master.h"
 #include "system/dungeon/dungeon-definition.h"
+#include "system/dungeon/quest-definition.h"
 #include "system/enums/dungeon/dungeon-id.h"
 #include "system/floor/floor-info.h"
 #include "system/floor/wilderness-grid.h"
