+++ src/floor/object-allocator.cpp
@@ -5,6 +5,7 @@
 #include "game-option/cheat-types.h"
 #include "grid/object-placer.h"
 #include "system/dungeon/dungeon-definition.h"
+#include "system/dungeon/quest-definition.h"
 #include "system/enums/terrain/terrain-tag.h"
 #include "system/floor/floor-info.h"
 #include "system/grid-type-definition.h"
