+++ src/floor/pattern-walk.cpp
@@ -19,6 +19,7 @@
 #include "status/bad-status-setter.h"
 #include "status/experience.h"
 #include "system/dungeon/dungeon-definition.h"
+#include "system/dungeon/quest-definition.h"
 #include "system/enums/dungeon/dungeon-id.h"
 #include "system/enums/terrain/pattern-tile-type.h"
 #include "system/enums/terrain/terrain-tag.h"
