+++ src/info-reader/general-parser.cpp
@@ -1,6 +1,5 @@
 #include "info-reader/general-parser.h"
 #include "artifact/fixed-art-types.h"
-#include "dungeon/quest.h"
 #include "info-reader/definition-hash-data.h"
 #include "info-reader/info-reader-util.h"
 #include "info-reader/parse-error-types.h"
@@ -14,6 +13,7 @@
 #include "system/baseitem/baseitem-definition.h"
 #include "system/baseitem/baseitem-list.h"
 #include "system/building-type-definition.h"
+#include "system/dungeon/quest-definition.h"
 #include "system/enums/terrain/terrain-tag.h"
 #include "system/floor/floor-info.h"
 #include "system/system-variables.h"
