+++ src/io-dump/character-dump.cpp
@@ -29,6 +29,7 @@
 #include "store/store.h"
 #include "system/angband-system.h"
 #include "system/building-type-definition.h"
+#include "system/dungeon/quest-definition.h"
 #include "system/floor/floor-info.h"
 #include "system/floor/town-list.h"
 #include "system/inner-game-data.h"
