+++ src/load/load-zangband.cpp
@@ -1,6 +1,5 @@
 #include "load/load-zangband.h"
 #include "avatar/avatar.h"
-#include "dungeon/quest.h"
 #include "floor/dungeon-feeling.h"
 #include "game-option/option-flags.h"
 #include "info-reader/fixed-map-parser.h"
@@ -21,6 +20,7 @@
 #include "spell/spells-status.h"
 #include "system/building-type-definition.h"
 #include "system/dungeon/dungeon-record.h"
+#include "system/dungeon/quest-definition.h"
 #include "system/enums/dungeon/dungeon-id.h"
 #include "system/floor/floor-info.h"
 #include "system/floor/town-records.h"
