+++ src/load/old/load-v1-7-0.cpp
@@ -1,8 +1,8 @@
 #include "load/old/load-v1-7-0.h"
-#include "dungeon/quest.h"
 #include "game-option/birth-options.h"
 #include "load/load-util.h"
 #include "load/old/load-v1-5-0.h"
+#include "system/dungeon/quest-definition.h"
 #include "system/enums/dungeon/dungeon-id.h"
 #include "system/floor/floor-info.h"
 #include "system/player-type-definition.h"
