+++ src/load/quest-loader.cpp
@@ -1,6 +1,5 @@
 #include "load/quest-loader.h"
 #include "artifact/fixed-art-types.h"
-#include "dungeon/quest.h"
 #include "load/angband-version-comparer.h"
 #include "load/load-util.h"
 #include "load/load-zangband.h"
@@ -8,6 +7,7 @@
 #include "object-enchant/trg-types.h"
 #include "system/angband-exceptions.h"
 #include "system/artifact/artifact-definition.h"
+#include "system/dungeon/quest-definition.h"
 #include "system/floor/floor-info.h"
 #include "system/floor/town-list.h"
 #include "system/monrace/monrace-definition.h"
