+++ src/main/angband-initializer.cpp
@@ -14,7 +14,6 @@
 
 #include "main/angband-initializer.h"
 #include "autopick/autopick-menu-data-table.h"
-#include "dungeon/quest.h"
 #include "floor/wild.h"
 #include "io/files-util.h"
 #include "io/read-pref-file.h"
@@ -23,6 +22,7 @@
 #include "main/info-initializer.h"
 #include "market/building-initializer.h"
 #include "system/angband-system.h"
+#include "system/dungeon/quest-definition.h"
 #include "system/floor/town-list.h"
 #include "system/monrace/monrace-list.h"
 #include "system/monrace/monrace-records.h"
@@ -33,6 +33,7 @@
 #include "term/term-color-types.h"
 #include "time.h"
 #include "util/angband-files.h"
+#include "view/display-messages.h"
 #include "world/world.h"
 
 /*!
@@ -222,7 +223,15 @@ void init_angband(PlayerType *player_ptr, bool no_term)
     init_buildings();
 
     init_note(_("[配列を初期化しています... (クエスト)]", "[Initializing arrays... (quests)]"));
-    QuestList::get_instance().initialize();
+    try {
+        QuestList::get_instance().initialize();
+    } catch (const std::runtime_error &r) {
+        std::stringstream ss;
+        ss << _("ファイル読み込みエラー: ", "File loading error: ") << r.what();
+        msg_print(ss.str());
+        msg_erase();
+        quit(_("クエスト初期化エラー", "Error of quests initializing"));
+    }
 
     init_note(_("[データの初期化中... (宝物庫)]", "[Initializing arrays... (vaults)]"));
     init_vaults_info();
