+++ src/market/building-quest.cpp
@@ -3,6 +3,7 @@
 #include "dungeon/quest.h"
 #include "info-reader/fixed-map-parser.h"
 #include "market/building-util.h"
+#include "system/dungeon/quest-definition.h"
 #include "system/enums/dungeon/dungeon-id.h"
 #include "system/floor/floor-info.h"
 #include "system/floor/wilderness-grid.h"
