+++ src/melee/melee-spell-util.cpp
@@ -1,9 +1,9 @@
 #include "melee/melee-spell-util.h"
 #include "dungeon/dungeon-flag-types.h"
-#include "dungeon/quest.h"
 #include "floor/geometry.h"
 #include "monster/monster-info.h"
 #include "system/dungeon/dungeon-definition.h"
+#include "system/dungeon/quest-definition.h"
 #include "system/floor/floor-info.h"
 #include "system/monrace/monrace-definition.h"
 #include "system/monster-entity.h"
