+++ src/mspell/mspell-attack.cpp
@@ -2,7 +2,6 @@
 #include "blue-magic/blue-magic-checker.h"
 #include "core/disturbance.h"
 #include "dungeon/dungeon-flag-types.h"
-#include "dungeon/quest.h"
 #include "monster-floor/monster-move.h"
 #include "monster-race/race-ability-mask.h"
 #include "monster/monster-describer.h"
@@ -25,6 +24,7 @@
 #include "spell-realm/spells-hex.h"
 #include "system/angband-system.h"
 #include "system/dungeon/dungeon-definition.h"
+#include "system/dungeon/quest-definition.h"
 #include "system/enums/monrace/monrace-id.h"
 #include "system/floor/floor-info.h"
 #include "system/monrace/monrace-definition.h"
