+++ src/player/player-damage.cpp
@@ -53,6 +53,7 @@
 #include "status/element-resistance.h"
 #include "system/building-type-definition.h"
 #include "system/dungeon/dungeon-definition.h"
+#include "system/dungeon/quest-definition.h"
 #include "system/floor/floor-info.h"
 #include "system/inner-game-data.h"
 #include "system/item/item-entity.h"
