+++ src/save/save.cpp
@@ -14,7 +14,6 @@
 #include "save/save.h"
 #include "artifact/fixed-art-types.h"
 #include "core/object-compressor.h"
-#include "dungeon/quest.h"
 #include "inventory/inventory-slot-types.h"
 #include "io/files-util.h"
 #include "io/report.h"
@@ -31,6 +30,7 @@
 #include "system/artifact/artifact-definition.h"
 #include "system/artifact/artifact-list.h"
 #include "system/artifact/artifact-record.h"
+#include "system/dungeon/quest-definition.h"
 #include "system/floor/floor-info.h"
 #include "system/floor/town-list.h"
 #include "system/floor/wilderness-grid.h"
