+++ src/spell-kind/spells-grid.cpp
@@ -6,6 +6,7 @@
 #include "game-option/birth-options.h"
 #include "grid/grid.h"
 #include "system/dungeon/dungeon-definition.h"
+#include "system/dungeon/quest-definition.h"
 #include "system/enums/terrain/terrain-tag.h"
 #include "system/floor/floor-info.h"
 #include "system/grid-type-definition.h"
