+++ src/wizard/wizard-game-modifier.cpp
@@ -10,6 +10,7 @@
 #include "io/input-key-requester.h"
 #include "player-info/self-info.h"
 #include "system/building-type-definition.h"
+#include "system/dungeon/quest-definition.h"
 #include "system/enums/dungeon/dungeon-id.h"
 #include "system/enums/monrace/monrace-id.h"
 #include "system/floor/floor-info.h"
